% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{calc_corr}
\alias{calc_corr}
\title{Estimate the sample correlation between GWAS datasets}
\usage{
calc_corr(data, cor_SNPs)
}
\arguments{
\item{data}{A list with following elements: SNP, beta_exp, se_exp, beta_out, se_out (typically returned by \code{format_data}).}

\item{cor_SNPs}{A vector of SNPs used to compute the correlation.}
}
\value{
Estimated sample correlation matrix
}
\description{
Estimate the sample correlation between GWAS datasets
}
