% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnps_new.R
\name{check_key}
\alias{check_key}
\title{Original check_key function from the rsnps R package}
\usage{
check_key(key = "")
}
\description{
Original check_key function from the rsnps R package
}
\keyword{internal}
