% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{extract_data}
\alias{extract_data}
\title{Extract selected SNPs from data}
\usage{
extract_data(files, SNP_list = NULL, message = TRUE)
}
\arguments{
\item{files}{A vector of file names.}

\item{SNP_list}{If not null, only returns SNPs in this list; otherwise returns the original dataset.}

\item{message}{Logical indicator for message printing.}
}
\value{
A list of data tables, each corresponding to one file.
}
\description{
Extract selected SNPs from data
}
