% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{filter_SNPs}
\alias{filter_SNPs}
\title{Filter SNPs for different purporses}
\usage{
filter_SNPs(
  sel_files = NULL,
  exp_files = NULL,
  out_files = NULL,
  SNP_inter = NULL,
  utility = c("inference", "marker_sel", "marker_exp", "correlation"),
  pval_thres = 5e-08,
  clump_r2 = switch(utility, inference = 0.01, marker_sel = 0.3, marker_exp = 0.3),
  plink_exe = NULL,
  plink_refdat = NULL
)
}
\arguments{
\item{sel_files}{A vector of GWAS summary data file names for selection}

\item{exp_files}{A vector of GWAS summary data file names for exposures}

\item{out_files}{A vector of GWAS summary data file names for outcomes}

\item{SNP_inter}{"Intersection" SNPs that are in all files returned by \code{get_SNP_intersection}.}

\item{utility}{Purpose of filtering SNPs.}

\item{pval_thres}{P-value threshold used for filtering SNPs (lower bound for \code{utility=correlation} and upper bound otherwise).}

\item{clump_r2}{PLINK argument clump-r2 (R-squared for clumping)}

\item{plink_exe}{Command to execute PLINK}
}
\value{
a vector of SNP names.
}
\description{
Filter SNPs for different purporses
}
