% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{filter_SNPs_basic}
\alias{filter_SNPs_basic}
\title{Select SNPs by p-value filtering and LD clumping}
\usage{
filter_SNPs_basic(
  files = NULL,
  SNP_inter = NULL,
  max_pval_files = files,
  max_pval_thres = 1,
  min_pval_files = files,
  min_pval_thres = 0,
  clump = FALSE,
  clump_pval_files = files,
  clump_r2 = 0.001,
  plink_exe = NULL,
  plink_refdat = NULL
)
}
\arguments{
\item{files}{A vector of file names}

\item{SNP_inter}{"Intersection" SNPs that are in all files returned by \code{get_SNP_intersection}.}

\item{max_pval_files}{Files used for upper-bounding p-values}

\item{max_pval_thres}{Threshold to upper-bound p-values.}

\item{min_pval_files}{Files used for lower-bounding p-values}

\item{min_pval_thres}{Threshold to lower-bound p-values.}

\item{clump}{Whether LD clumping should be performed}

\item{clump_pval_files}{Files to be used to compute minimum p-value for clumping.}

\item{clump_r2}{PLINK argument clump-r2 (R-squared for clumping)}

\item{plink_exe}{Command to execute PLINK}
}
\description{
Select SNPs by p-value filtering and LD clumping
}
\details{
When upper-bounding the p-values, only one p-value needs to be below \code{max_pval_thres}. When lower-bounding the p-values, all p-values need to be above \code{min_pval_thres}.
}
\keyword{internal}
