% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModes.R
\name{findLocalModes}
\alias{findLocalModes}
\title{findLocal maximum modes of a series of points}
\usage{
findLocalModes(v, npts = 21)
}
\description{
findLocal maximum modes of a series of points
}
\keyword{internal}
