% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModes.R
\name{findModes}
\alias{findModes}
\title{Use the multiple modes of the robust profile likelihood function to find out multiple pathways in MR and their marker SNPs.}
\usage{
findModes(b_exp, b_out, se_exp, se_out, marker.data = NULL,
  mode.lmts = c(-5, 5), cor.mat = NULL, loss.function = c("tukey",
  "huber", "l2"), k.findmodes = switch(loss.function[1], l2 = NA, huber =
  1.345, tukey = 3), beta.mode = NULL, include.thres = 1,
  exclude.thres = 2, map.marker = T, ldThres = 0.9,
  npoints = 10000)
}
\arguments{
\item{b_exp}{A matrix of size \code{p * k} for the effect sizes of \code{p} number of selected independent SNPs (instruments) on \code{k} risk factors. SNPs should be provided after clumping.}

\item{b_out}{A vector of length \code{p} for the effect sizes of the selected \code{p} SNPs on the disease (outcome)}

\item{se_exp}{A matrix of size \code{p * k} for the standard deviations of each entry in \code{b_exp}}

\item{se_out}{A vector of length \code{p} for the standard deviations of each element in \code{b_out}}

\item{marker.data}{A list of the data used to find markers. Default is NULL, which is using the same sets of SNPs as find the modes for finding the markers. Another choice is to pass the output \code{marer.data} from the grapple function \code{getInput}. If the user wants to provide his/her own list, then it should be a list containing 4 elements, \code{b_exp}, \code{b_out}, \code{se_exp} and \code{se_out}.}

\item{cor.mat}{Either NULL or a \code{k + 1} by \code{k + 1} symmetric matrix. The shared correlation matrix for \code{(b_exp[j], b_out[j])} across SNP j. Default is NULL, for the identity matrix}

\item{loss.function}{Loss function used, one of "tukey", "huber" or "l2". Default is "tukey", which is robust to outlier SNPs with large pleiotropic effects}

\item{k.findmodes}{Tuning parameters of the loss function, for loss "l2", it is NA, for loss "huber", default is 1.345 and for loss "tukey", default is 3.}

\item{beta.mode}{Allow providing values of the modes to find out marker SNPs for any given modes}

\item{include.thres}{Absolute value upper threshold of the standardized test statistics of one SNP on one mode for the SNP to be included as a marker for that mode, default is 1.4}

\item{exclude.thres}{Absolute value lower threshold of the standardized test statistics of one SNP on other modes for the SNP to be included as a marker for that mode, default is \code{qnorm(0.975)}}

\item{map.marker}{Whether map each marker to the earist gene or not. Default is TRUE if multiple markers are found. It is always FALSE if there is just one mode.}

\item{ldThres}{the parameter passed to the \code{queryhaploReg} function. Increase to 1 when there is a "timeout" error.}

\item{mode_lmts}{The range of \code{beta} that the modes are searched from. Default is \code{c(-5, 5)}}

\item{npoints:}{number of equally spaced points chosen for grid search of modes within the range \code{mode.lmts}.}
}
\value{
A list containing the following elements:
\item{fun}{}
\item{modes}{}
\item{p}{}
\item{markers}{}
\item{raw.modes}{}
\item{supp_gwas}{}
}
\description{
THis function can be run only for \code{k = 1} when there is only one risk factor
}
