% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{format_data}
\alias{format_data}
\title{Format GWAS data}
\usage{
format_data(dat_list, sel_files = NULL, exp_files = NULL, out_files = NULL)
}
\arguments{
\item{dat_list}{A list of GWAS data tables (typically returned by \code{harmonise_data_list}).}

\item{sel_files}{A vector of GWAS summary data file names for selection}

\item{exp_files}{A vector of GWAS summary data file names for exposures}

\item{out_files}{A vector of GWAS summary data file names for outcomes}
}
\value{
A list with the following variables:
\itemize{
\item \code{meta_data} (a data frame);
\item \code{beta_sel}, \code{se_sel}, \code{pval_sel} (returned if \code{sel_files} is not NULL);
\item \code{beta_exp}, \code{se_exp}, \code{pval_exp} (returned if \code{exp_files} is not NULL);
\item \code{beta_out}, \code{se_out}, \code{pval_out} (returned if \code{out_files} is not NULL);
\item \code{beta}, \code{se}, \code{pval} (returned if all files are NULL);
}
}
\description{
Group the same columns in a list of GWAS data tables together.
}
