% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{get_snp_intersection}
\alias{get_snp_intersection}
\title{Get intersection of the SNPs in GWAS summary datasets}
\usage{
get_snp_intersection(files, message = TRUE)
}
\description{
Get intersection of the SNPs in GWAS summary datasets
}
\details{
The first file is the selection dataset.
}
