% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{harmonise_data_list}
\alias{harmonise_data_list}
\title{Harmonise a list of GWAS summary data tables}
\usage{
harmonise_data_list(dat_list, fast = T)
}
\arguments{
\item{fast}{Whether to use a fast implementation of GWAS data harmonisation.}

\item{dat}{A list of \code{data.table} objects containing GWAS summary data (typically returned by \code{extract_data})}
}
\description{
Harmonise a list of GWAS summary data tables
}
\details{
When \code{fast} is \code{TRUE}, this function uses a modification of the \emph{TwoSampleMR::harmonise_data} that skips certain checks.
}
