% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{load_R_data}
\alias{load_R_data}
\title{Loads an R data file, and returns it}
\usage{
load_R_data(file)
}
\arguments{
\item{file}{Name of the R data file (.rda or .RData).}
}
\value{
First object in the R data file.
}
\description{
Loads an R data file, and returns it
}
\keyword{internal}
