% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnps_new.R
\name{ncbi_snp_summary}
\alias{ncbi_snp_summary}
\title{Query NCBI's dbSNP for summary information on a set of SNPs. This function is modified from \code{rsnps} package to make it work with the NCBI data format change}
\usage{
ncbi_snp_summary(x, key = NULL, ...)
}
\arguments{
\item{x}{A vector of SNPs (with or without 'rs' prefix)}

\item{key}{(character) NCBI Entrez API key. optional. 
See "NCBI Authenication" in [rsnps-package]}

\item{...}{Curl options passed on to [crul::HttpClient]}
}
\value{
data.frame with three columns: \code{snp_id}, \code{gene2} and \code{chrpos}. SNPs not found are omitted
}
\description{
Query NCBI's dbSNP for summary information on a set of SNPs. This function is modified from \code{rsnps} package to make it work with the NCBI data format change
}
