% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{plink_clump}
\alias{plink_clump}
\title{Wrapper function to call PLINK}
\usage{
plink_clump(
  dat,
  plink_exe,
  plink_refdat,
  clump_kb = 1000,
  clump_r2 = 0.05,
  clump_p1 = 1,
  clump_p2 = 1,
  tempdir = "temp"
)
}
\arguments{
\item{dat}{A data frame containing SNPs and p-values}

\item{plink_exe}{Command to execute PLINK}

\item{clump_kb}{PLINK argument clump-kb}

\item{clump_r2}{PLINK argument clump-r2 (R-squared for clumping)}

\item{clump_p1}{PLINK argument clump-p1}

\item{clump_p2}{PLINK argument clump-p2}

\item{tempdir}{Temporary directory to save the output of PLINK}

\item{refdat}{File path to reference genome data}
}
\description{
Wrapper function to call PLINK
}
\keyword{internal}
