% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_fun_general.R
\name{rho.tukey}
\alias{rho.tukey}
\title{Tukey's beweight loss function and its derivatives}
\usage{
rho.tukey(r, k = 4.685, deriv = 0)
}
\description{
Tukey's beweight loss function and its derivatives
}
