% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{extract_data}
\alias{extract_data}
\title{Extract selected SNPs from data}
\usage{
extract_data(files, SNP_list = NULL, message = TRUE)
}
\description{
Extract selected SNPs from data
}
\examples{
files <- file.path("../data", list.files("../data"))
tmp <- get_snp_intersection(files)
tmp2 <- plink_clump(tmp, "../util/plink_mac/plink", refdat = "../util/data_maf0.01_rs")
dat <- extract_data(files, tmp2$SNP)
dat <- harmonise_data_list(dat)

}
