% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModes.R
\name{findModes}
\alias{findModes}
\title{Use the multiple modes of the robust profile likelihood function to find out multiple pathways in MR and their marker SNPs.}
\usage{
findModes(b_exp, b_out, se_exp, se_out, b_exp_st = b_exp,
  b_out_st = b_out, se_exp_st = se_exp, se_out_st = se_out,
  mode.lmts = c(-2, 2), cor.mat = NULL, loss.function = c("tukey",
  "l2", "huber"), k = switch(loss.function[1], l2 = NA, huber = 1.345,
  tukey = 3), beta.mode = NULL, include.thres = 1.4,
  exclude.thres = qnorm(0.975), map.marker = T)
}
\arguments{
\item{b_exp}{A vector of length \code{p} for the effect sizes of \code{p} number of independent SNPs on the risk factor}

\item{b_out}{A vector of length \code{p} for the effect sizes of the \code{p} SNPs on the outcome}

\item{se_exp}{A vector of length \code{p} for the standard deviations of the effect sizes in \code{b_exp}}

\item{se_out}{A vector of length \code{p} for the standard deviations of the effect sizes in \code{b_out}}

\item{b_exp_st}{A vector of length \code{q} for the effect sizes of \code{q} candidates SNPs on the risk factor where the marker SNPs are selected from. Default is the same as \code{b_exp}}

\item{b_out_st}{A vector of length \code{q} for the effect sizes of the \code{q} same list of SNPs as \code{b_exp_st} on the outcome, default is the same as \code{b_out}}

\item{se_exp_st}{A vector of length \code{q} for the standard deviations of the effect sizes in \code{b_exp_st}, default is the same as \code{se_exp}}

\item{se_out_st}{A vector of length \code{q} for the standard deviations of the effect sizes in \code{b_out_st}, default is the same as \code{se_out}}

\item{cor.mat}{Either NULL or a 2 by 2 symmetric matrix. The correlation matrix of estimated effect sizes on the isk factor and the outcome. Default is NULL, for the identity matrix}

\item{loss.function}{Loss function used, one of "l2", "huber" and "tukey". Default is "tukey". One should use "tukey" or "huber" in order to find multiple modes}

\item{k}{Tuning parameters of the loss function, for loss "l2", it is NA, for loss "huber", default is 1.345 and for loss "tukey", default is 3.}

\item{beta.mode}{Allow providing values of the modes to find out marker SNPs for any given modes}

\item{include.thres}{Absolute value upper threshold of the standardized test statistics of one SNP on one mode for the SNP to be included as a marker for that mode, default is 1.4}

\item{exclude.thres}{Absolute value lower threshold of the standardized test statistics of one SNP on other modes for the SNP to be included as a marker for that mode, default is \code{qnorm(0.975)}}

\item{map.marker}{Whether map each marker to the earist gene or not. Default is TRUE if multiple markers are found. It is always FALSE if there is just one mode.}

\item{mode_lmts}{The range of \code{beta} that the modes are searched from. Default is \code{c(-2, 2)}}
}
\value{
A list of the modes, RAP plot and the markers
}
\description{
Use the multiple modes of the robust profile likelihood function to find out multiple pathways in MR and their marker SNPs.
}
