% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModes.R
\name{findModes}
\alias{findModes}
\title{Use the multiple modes of the robust profile likelihood function to find out multiple pathways in MR and their marker SNPs.}
\usage{
findModes(data, p.thres = NULL, marker.data = NULL,
  marker.p.thres = NULL, mode.lmts = c(-5, 5), cor.mat = NULL,
  loss.function = c("tukey", "huber", "l2"),
  k.findmodes = switch(loss.function[1], l2 = NA, huber = 1.345, tukey =
  3), include.thres = 1, exclude.thres = 2, map.marker = T,
  ldThres = 0.9, npoints = 10000)
}
\arguments{
\item{data}{A data frame containing the information of the selected genetic instruments. 
One can simply take the \code{data} element from the output of function \code{getInput}, or provide their own data frame.
The required columns include \code{SNP} for the SNP rsID, the columns \code{gamma_exp1} to \code{gamma_expk} for the 
estimated effect sizes of the SNPs on risk factors \code{1} to \code{k}, the columns \code{se_exp1} to \code{se_expk} for their
standard deviations, the columns \code{gamma_out1} to \code{gamma_outm} for the 
estimated effect sizes of the SNPs on diseases \code{1} to \code{m},the columns \code{se_out1} to \code{se_outm} for their 
standard deviations.}

\item{p.thres}{The p-value threshold for SNP selection. The SNPs whose \code{selection_pvals} are less than \code{p.thres} are selected. 
The default value is NULL, which is to include all SNPs in \code{data}. If it is not NULL, then \code{data} should have a column \code{selection_pvals}
that stores the selection p-value of each SNP.}

\item{marker.data}{A data frame containing the information of candidate marker genes.
Default is NULL, which sets \code{marker.data} to \code{data}. 
Another choice is to use the \code{marker.data} element in the output of \code{getInput}.}

\item{marker.p.thres}{P-value threshold for marker SNP selection. See \code{p.thres}}

\item{cor.mat}{Either NULL or a \code{k + 1} by \code{k + 1} symmetric matrix. The shared correlation matrix for \code{(b_exp[j], b_out[j])} across SNP j. Used only when \code{input.list} is NULL and the default value is NULL, for the identity matrix.}

\item{loss.function}{Loss function used, one of "tukey", "huber" or "l2". Default is "tukey", which is robust to outlier SNPs with large pleiotropic effects}

\item{k.findmodes}{Tuning parameters of the loss function, for loss "l2", it is NA, for loss "huber", default is 1.345 and for loss "tukey", default is 3.}

\item{include.thres}{Absolute value upper threshold of the standardized test statistics of one SNP on one mode for the SNP to be included as a marker for that mode, default is 1.4}

\item{exclude.thres}{Absolute value lower threshold of the standardized test statistics of one SNP on other modes for the SNP to be included as a marker for that mode, default is \code{qnorm(0.975)}}

\item{map.marker}{Whether map each marker to the earist gene or not. Default is TRUE if multiple markers are found. It is always FALSE if there is just one mode.}

\item{ldThres}{the parameter passed to the \code{queryhaploReg} function. Increase to 1 when there is a "timeout" error.}

\item{npoints}{Number of equally spaced points chosen for grid search of modes within the range \code{mode.lmts}.}

\item{mode_lmts}{The range of \code{beta} that the modes are searched from. Default is \code{c(-5, 5)}}
}
\value{
A list containing the following elements:
\item{fun}{The profile likelihood function with argument \code{beta}}
\item{modes}{The position of modes. Only include modes where marker genes can be detected}
\item{p}{The profile likelihood plot with gene markers when there are multiple modes. The range of the x.axis depends on the distance between the maximum mode and minimum mode when there are multiple modes.}
\item{markers}{A data frame of marker information}
\item{raw.modes}{All modes of the profile likelihood function within the range of \code{mode_lmts}}
\item{supp_gwas}{More information about the markers.}
}
\description{
THis function can be run only for \code{k = 1} when there is only one risk factor
}
