% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_new.R
\name{get_SNP_intersection}
\alias{get_SNP_intersection}
\title{Get intersection of the SNPs in several GWAS summary datasets}
\usage{
get_SNP_intersection(files, message = TRUE)
}
\arguments{
\item{files}{A vector of file names}

\item{message}{Logical indicator for message printing}
}
\value{
A \code{data.table} object with the SNP name and p-values in all summary datasets.
}
\description{
Get intersection of the SNPs in several GWAS summary datasets
}
