% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getInput}
\alias{getInput}
\title{Preprocess GWAS summary statistics datasets}
\usage{
getInput(sel.files, exp.files, out.file, plink_exe, plink_refdat,
  max.p.thres = 0.01, cal.cor = T, p.thres.cor = 0.5,
  get.marker.candidates = T, sample.split.marker.candidates = F,
  marker.p.thres = 1e-05, clump_r2 = 0.001,
  clump_r2_formarkers = 0.05)
}
\arguments{
\item{sel.files}{A vector of length \code{k} of the GWAS summary statistics file names of the \code{k} risk factors SNP selection. Each GWAS file is a ".csv" or ".txt" file containing a data frame that at least has a column "SNP" for the SNP ids, "pval" for the p-values.}

\item{exp.files}{A vector of length \code{k} of the GWAS summary statistics file names of the \code{k} risk factors for getting the effect sizes and standard deviations. Each GWAS file should have a column "SNP" for the SNP ids, "beta" for the effect sizes, "se" for the standard deviation, "effect_allele" for the effect allele of the SNP (capitalized letters) and "other_allele" for the other allele of the SNP (capitalized letters).}

\item{out.file}{The GWAS summary statistics file name for the outcome data. Each GWAS file should have a column "SNP" for the SNP ids, "beta" for the effect sizes, "se" for the standard deviation, "effect_allele" for the effect allele of the SNP (capitalized letters) and "other_allele" for the other allele of the SNP (capitalized letters).}

\item{plink_exe}{The executable file of PLINK. PLINK should be first downloaded from \url{https://www.cog-genomics.org/plink2}.}

\item{plink_refdat}{The reference genotype files (.bed, .bim, .fam) for clumping using PLINK (loaded with --bfile).}

\item{max.p.thres}{The upper threshold of the selection p-values for a SNP to be selected before clumping. It only requires that at least one of the p-values of the risk factors of the SNPs to be below the threshold. Default is \code{0.01}.}

\item{cal.cor}{Whether calculate the \code{(k + 1)} by \code{(k + 1)} correlation matrix between the \code{k} risk factors and the outcome. The default is TRUE}

\item{p.thres.cor}{The lower threshold of the p-values for a SNP to be used in calculating the correlation matrix}

\item{get.marker.candidates}{Whether getting SNPs which are used for mode marker selection. Only applies to cases where the number of risk factors \code{k = 1}. Default is TRUE for \code{k = 1}.}

\item{marker.p.thres}{Selection threshold of p-values for mode markers. Default is \code{1e-5}.}

\item{clump_r2}{The clumping r2 threshold in PLINK for genetic instrument selection. Default is set to 0.001 for selection of independent SNPs.}

\item{clump_r2_formarkers}{The clumping r2 threshold in PLINK. Default is set to 0.05 for selection of candidates for the marker SNPs.}
}
\description{
This function has GWAS summary statistics data files as inputs, perform genetic instrument selection and return matrices that are ready to use for GRAPPLE
}
