% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_fun_general.R
\name{grappleRobustEst}
\alias{grappleRobustEst}
\title{Robust multivariate MR estimation}
\usage{
grappleRobustEst(b_exp, b_out, se_exp, se_out, cor.mat = NULL,
  loss.function = c("l2", "huber", "tukey"), k = switch(loss.function[1], l2
  = NA, huber = 1.345, tukey = 4.685), suppress.warning = FALSE,
  diagnosis = FALSE, niter = 20, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{b_exp}{A matrix of size \code{p * k} for the effect sizes of \code{p} number of independent SNPs on \code{k} risk factors}

\item{b_out}{A vector of length \code{p} for the effect sizes of the \code{p} SNPs on the outcome}

\item{se_exp}{A matrix of size \code{p * k} for the standard deviations of the effect sizes in \code{b_exp}}

\item{se_out}{A vector of length \code{p} for the standard deviations of the effect sizes in \code{b_out}}

\item{cor.mat}{Either NULL or a \code{k + 1} by \code{k + 1} symmetric matrix. The correlation matrix of estimated effect sizes on the \code{k} risk factors and the outcome. Default is NULL, for the identity matrix}

\item{loss.function}{Loss function used, one of "l2", "huber" and "tukey"}

\item{k}{Tuning parameters of the loss function, for loss "l2", it is NA, for loss "huber", default is 1.345 and for loss "tukey", default is 4.685}

\item{suppress.warning}{Whether suppress warning messages or not, default is FALSE}

\item{diagnosis}{Run diagnosis analysis based on the residuals or not, default is FALSE}

\item{niter}{Number of iterations for optimization. Default is 20}

\item{tol}{Tolerance for convergence, default is the square root of the smallest positive floating number depending on the machine R is running on}
}
\description{
Robust multivariate MR estimation
}
